Replace `int` ids to `long` type

+++ data/src/main/java/telegram4j/json/ContactData.java
@@ -26,7 +26,7 @@ public interface ContactData {
     Optional<String> lastName();
 
     @JsonProperty("user_id")
-    Optional<Integer> userId();
+    Optional<Long> userId();
 
     Optional<String> vcard();
 }
